refactoring for readability

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: gdelhota <gdelhota@student.42perpigna      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/14 17:19:19 by gdelhota          #+#    #+#             */
-/*   Updated: 2024/11/28 02:14:45 by gdelhota         ###   ########.fr       */
+/*   Updated: 2024/11/28 05:06:20 by gdelhota         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,11 +25,11 @@
 #  define BUFFER_SIZE 32
 # endif
 
-int		trim_endline(char **s, char *buffer, size_t size);
-char	*strn_append(char *s1, char *s2, size_t n, char sadface);
-char	*ft_strndup(char *s, size_t n);
+int		trim_endline(char **s, char *buffer, int size);
+char	*strn_append(char *s1, char *s2, int n, char sadface);
+char	*ft_strndup(char *s, int n);
 void	ft_strncpy(char *src, char *dst, int n);
-size_t	ft_strlen(char *s);
+int		ft_strlen(char *s);
 char	*get_next_line(int fd);
 
 #endif
